+++ Subsurface-mobile.pro
@@ -28,7 +28,6 @@ SOURCES += subsurface-mobile-main.cpp \
 	core/qtserialbluetooth.cpp \
 	core/plannernotes.c \
 	core/uemis-downloader.c \
-	core/applicationstate.cpp \
 	core/qthelper.cpp \
 	core/checkcloudconnection.cpp \
 	core/color.cpp \
